Form Printing

+++ CourtApp.Web/Areas/Client/Model/ClientFeeViewModel.cs
@@ -4,7 +4,7 @@ namespace CourtApp.Web.Areas.Client.Model
 {
     public class ClientFeeViewModel
     {
-        public float FeeSettled { get; set; }
-        public float FeeAdvance { get; set; }
+        public Decimal FeeSettled { get; set; }
+        public Decimal FeeAdvance { get; set; }
     }
 }
